backend/pyproject.toml
[project]
name = "auction-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "pydantic>=2.8",
  "pydantic-settings>=2.3",
  "python-dotenv>=1.0",
  "python-socketio[asgi]>=5.11"
]

[tool.uvicorn]
factory = true
host = "0.0.0.0"
port = 8000
reload = true
